+++ src/Ozon.Route256.Five.OrderService/Contracts/GetStatus/GetStatusRequest.cs
@@ -1,3 +1,3 @@
 namespace Ozon.Route256.Five.OrderService.Contracts.GetStatus;
 
-public record GetStatusRequest(int Id);
+public record GetStatusRequest(long Id);
