+++ src/Ozon.Route256.Five.OrderService/Features/CancelOrder/CancelOrderCommand.cs
@@ -2,4 +2,4 @@ using Ozon.Route256.Five.OrderService.Cqrs;
 
 namespace Ozon.Route256.Five.OrderService.Features.CancelOrder;
 
-public record CancelOrderCommand(int Id) : ICommand;
+public record CancelOrderCommand(long Id) : ICommand;
