+++ src/Ozon.Route256.Five.OrderService/Features/GetOrderStatus/GetOrderStatusQuery.cs
@@ -4,4 +4,4 @@ using Ozon.Route256.Five.OrderService.Cqrs;
 
 namespace Ozon.Route256.Five.OrderService.Features.GetOrderStatus;
 
-public record GetOrderStatusQuery(int Id) : GetStatusRequest(Id), IQuery<GetStatusResponse>;
+public record GetOrderStatusQuery(long Id) : GetStatusRequest(Id), IQuery<GetStatusResponse>;
